loads of updates

+++ PowerLifting.Entity/ProgramLogs/DTO/ProgramLogRepSchemeDTO.cs
@@ -6,7 +6,7 @@
         public int ProgramLogExerciseId { get; set; }
         public int SetNo { get; set; }
         public int NoOfReps { get; set; }
-        public double WeightLifted { get; set; }
+        public decimal WeightLifted { get; set; }
         public double? Percentage { get; set; }
         public string Comment { get; set; }
         public bool AMRAP { get; set; } //As many reps as possible for this set
